friendly log support can provide a function to derive an anno
Make it possible to provide a function to derive an annotation when one
is omitted.  In deuxdrop this is being used to generate a call stack,
then skip all the Q and node related frames to find the first
interesting one and use that to generate a "func @ file:line" style
string.  (I was previously just providing the whole stack which required
clicking into it and then human introspection, which is a waste.)

I also made us provide some explicit Q annotations for things where V8's
guesses were wrong or confusing.  For example, 'fail' was labeling some
Q.all internals in a confusing way.